clan-points-bot/requirements.txt
discord.py
python-dotenv
watchdog
pymongo
aiohttp
gspread
gspread_formatting